fix: updated user_roles to have created_at column

+++ backend/src/TeleHealth.Api/Domain/Entities/UserRole.cs
@@ -4,7 +4,7 @@ namespace TeleHealth.Api.Domain.Entities;
 
 public sealed class UserRole
 {
-    public int UserId { get; init; }
+    public long UserId { get; init; }
     public int RoleId { get; init; }
     public Instant CreatedAt { get; set; }
 }
